+++ DGDownloadManager/DGDownloadManagerFile.m
@@ -215,7 +215,7 @@
 {
     if ([response isKindOfClass:NSHTTPURLResponse.class])
     {
-        int statusCode = [(NSHTTPURLResponse *)response statusCode];
+        NSInteger statusCode = [(NSHTTPURLResponse *)response statusCode];
         if (statusCode >= 400)
         {
             [connection cancel];
